Move some types to types/
